pyproject.toml
[tool.poetry]
name = "api-python-flask-template"
version = "0.1.0"
description = ""
authors = ["Developer Name <developer@rapidinnovation.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
flask-restx = "^1.3.0"
flask-pymongo = "^2.3.0"
pyjwt = "^2.8.0"
redis = "^5.0.1"
cryptography = "^42.0.2"
pika = "^1.3.2"
celery = "^5.3.6"
google-cloud-storage = "^2.14.0"
flask-cors = "^4.0.0"
pandas = "^2.2.0"
pydantic = "^2.6.4"
sentry-sdk = "^1.45.0"
gunicorn = "^22.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "^24.1.1"
isort = "^5.13.2"
pytest = "^8.0.0"
pytest-flask = "^1.3.0"
pika-stubs = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
